Force close application after failed test
Force close is necessary because app.Quit() sometimes does not work
Also move application restart from the test task to the main thread

The tasks do not stop after the timeout, and there is no way to
cancel them if they are hung in third party code (like opening
the MSO application). The only way is to kill the application's
process, then they will throw an exception and continue.