Get PDF from Collabora Online (if not exist)
TODO: track time statistics

This patch requires that the converted file name
be calculated before being passed to ConvertFile,
instead of letting ConvertFile devise its own scheme.

This allows checking if the file exists
before requesting another PDF.
The same logic COULD be used to avoid converting to MSO format.

I'm not sure if this is completely necessary,
but it allows mso-test to run more frequently
without affecting the PDF testing
(which requires the use of a companion tool).