diff-pdf-page-statistics.py: exclude explicit files from testing
These files have proven to be false positives
and are excluded because they seem to have
no redeeming elements that make them worthy
of being considered for any reason.

True duplicates are thought to be already be removed,
but these are probably resaved, or historical versions
that are basically identical for the purposes of this test.