mso-test: force using Word 2013/compat15 for convert-to docx
Unfortunately, this only works for soffice.
I wasn't able to do the same thing for cool.
Certainly it would cause complications for onlyOffice, etc. anyway.

There is a PDFVer parameter that MIGHT suggest the way forward,
but it didn't work out of the box since it errors
unless a valid pdf version is provided.
It looks like we COULD force an option
/convert-to/docx?options=xyz
which is just passed verbatim to the process.

I noticed that our "download" and "save as" methods
also result in Word 2007 format...