mso-test: force using Word 2013/compat15 for convert-to docx
The issue is that our conversion does NOT
attempt to keep the same compat level,
but simply does a DOCX export with the 'preferred' filter.

Unfortunately, this can only be forced for soffice.exe
I wasn't able to do the same thing for COOL.
Certainly it would cause complications for onlyOffice, etc. anyway.

So why worry about it?
Well, this is a comparison tool,
reporting on differences with a previous run.
So, for the export comparison,
both runs need to be exporting with the same filter.

For command-line conversion, DOCX was exporting as Word 2007
until 25.8's early commit c90c1d4435a64531e1d9b41d0a8dc0b91ba236cd.