diff-pdf-page-statistics.py: exit if binary format missed PDF rename
There are quite a few manual steps in this process.
One of them is to rename the export (round-tripped) PDF
to match the base name, when binary files were round-tripped
into their respective modern XML formats.

Otherwise we risk spending 2 days processing a comparison
only to find out that the history PDF is always missing.
So this is just a sanity check helper.