diff-pdf-page-statistics.py: mostly remove stray red pixels
I believe the problem mostly arose from quantize(2)
forcing nearly-transparent pixels to red,
while those colors became transparent in the grayscale layer.

In any case, this moves ahead with a TODO in the code,
which was to remove the reliance on identifying the RED pixel.
I MUST remove that if i get rid of quantize(2).