Bug fixing, comments, and MPLv2 headers
One bug was that if the bmp was larger than the other it would do funny things.
Changed the indexing for bmps's so they are custom to each other, not based off
min_width. Cleaned some the pixel code up with.

Added case checking into main.

Adding comments to the files which should be ignored by git.

Signed-off-by: Elliot Freebairn <elliot.freebairn@collabora.com>